feature: StationShop.sellInventory()

+++ src/main/java/io/solar/dto/GoodsDto.java
@@ -8,5 +8,5 @@ public class GoodsDto {
     private Long station;
     private Long product;
     private Long amount;
-    private Float price;
+    private Long price;
 }
